Complete

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
@@ -52,13 +53,35 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle file upload
+        if ($request->hasFile('cover_image')) {
+            //Get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //Get just Filename
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+
+            //Get just Extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //File Name oto store
+            $fileNameToStore = $fileName . '_' . time() . '.' . $extension;
+
+            //Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         //Create Post
         $post = new Post;
         $post->title = $request->title;
         $post->body = $request->body;
+        $post->cover_image = $fileNameToStore;
         $post->user_id = auth()->user()->id;
         $post->save();
 
@@ -108,10 +131,31 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+        //Handle file upload
+        if ($request->hasFile('cover_image')) {
+            //Get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //Get just Filename
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+
+            //Get just Extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //File Name oto store
+            $fileNameToStore = $fileName . '_' . time() . '.' . $extension;
+
+            //Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         //Create Post
         $post = Post::find($id);
         $post->title = $request->title;
         $post->body = $request->body;
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Your Post Has Been Updated');
@@ -126,6 +170,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (auth()->user()->id != $post->user_id) {
+            return abort(401);
+            // return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        if ($post->cover_image != 'noimage.jpg') {
+            // Delete Image
+            Storage::delete('public/cover_images/' . $post->cover_image);
+        }
         $post->delete();
 
         return redirect('/posts')->with('error', 'Your Post Has Beed Deleted');
